commit 0.3.1 //modify user module add userlogbl&userlogservice
