+++ HBMLClient/src/main/java/businesslogic/userbl/UserStaffImpl.java
@@ -44,22 +44,4 @@ public class UserStaffImpl implements UserStaffService{
 
     }
 
-    @Override
-    public ResultMessage login(String accountName, String pwd) throws Exception{
-
-        return userHelper.login(accountName,pwd);
-
-    }
-
-    @Override
-    public ResultMessage signup(UserVO vo) throws Exception{
-
-        return userHelper.signup(vo);
-
-    }
-
-    @Override
-    public ResultMessage logout(String accountName) throws Exception {
-        return userHelper.logout(accountName);
-    }
 }
