+++ HBMLClient/src/test/test_Alex_BL.java
@@ -9,7 +9,7 @@ public class test_Alex_BL {
     static test_Alex_Promotion_BL test_alex_promotion_bl=new test_Alex_Promotion_BL();
 
     static void test_User_BL()throws Exception{
-        //test_alex_user_bl.signup();
+        test_alex_user_bl.signup();
         //test_alex_user_bl.login();
         //test_alex_user_bl.getUserInfo();
         //test_alex_user_bl.delete();
